+++ packages/cli/CHANGELOG.md
@@ -1,5 +1,134 @@
 # lingo.dev
 
+## 0.93.13
+
+### Patch Changes
+
+- Updated dependencies [[`1bccb7e`](https://github.com/lingodotdev/lingo.dev/commit/1bccb7ed51ac1f13ea79e618bbee551d5529efdc)]:
+  - @lingo.dev/_compiler@0.1.8
+
+## 0.93.12
+
+### Patch Changes
+
+- Updated dependencies [[`5b68641`](https://github.com/lingodotdev/lingo.dev/commit/5b686414f363f8ee4b79fd4e804a434db5cfcb36)]:
+  - @lingo.dev/_compiler@0.1.7
+
+## 0.93.11
+
+### Patch Changes
+
+- Updated dependencies [[`1f9db11`](https://github.com/lingodotdev/lingo.dev/commit/1f9db11a53d8c75ce0e83517b73d43544d0f0fd2)]:
+  - @lingo.dev/_react@0.2.1
+
+## 0.93.10
+
+### Patch Changes
+
+- Updated dependencies [[`7a5898b`](https://github.com/lingodotdev/lingo.dev/commit/7a5898b12dcd0015a5e57236bf65172cedb8a6ee)]:
+  - @lingo.dev/_compiler@0.1.6
+
+## 0.93.9
+
+### Patch Changes
+
+- Updated dependencies [[`7013b53`](https://github.com/lingodotdev/lingo.dev/commit/7013b5300d6c2c26f39da62b5ad2c7cf11158c74)]:
+  - @lingo.dev/_compiler@0.1.5
+
+## 0.93.8
+
+### Patch Changes
+
+- [#853](https://github.com/lingodotdev/lingo.dev/pull/853) [`cb7d5e2`](https://github.com/lingodotdev/lingo.dev/commit/cb7d5e213282c00af658159472183a763f84ca3d) Thanks [@vrcprl](https://github.com/vrcprl)! - Fix groq api key retrieval from .env
+
+- Updated dependencies [[`cb7d5e2`](https://github.com/lingodotdev/lingo.dev/commit/cb7d5e213282c00af658159472183a763f84ca3d)]:
+  - @lingo.dev/_compiler@0.1.4
+
+## 0.93.7
+
+### Patch Changes
+
+- [`5d27455`](https://github.com/lingodotdev/lingo.dev/commit/5d2745545044cbaddb099f7920c96fe198879ba3) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - add boolean parsing
+
+## 0.93.6
+
+### Patch Changes
+
+- [#843](https://github.com/lingodotdev/lingo.dev/pull/843) [`b67a331`](https://github.com/lingodotdev/lingo.dev/commit/b67a33141253fa755b5531e52cd690bf5824d4b6) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - explicit source / targets in run cmd, parallel mode for ci/cd
+
+## 0.93.5
+
+### Patch Changes
+
+- Updated dependencies [[`e75e615`](https://github.com/lingodotdev/lingo.dev/commit/e75e615ab17e279deb5a505dbda682fdfc7ead62)]:
+  - @lingo.dev/_react@0.2.0
+
+## 0.93.4
+
+### Patch Changes
+
+- [`f42cff8`](https://github.com/lingodotdev/lingo.dev/commit/f42cff8355b1ff7bba1445bd04d11ee4672903c2) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - flat reexports
+
+- Updated dependencies [[`f42cff8`](https://github.com/lingodotdev/lingo.dev/commit/f42cff8355b1ff7bba1445bd04d11ee4672903c2)]:
+  - @lingo.dev/_compiler@0.1.3
+
+## 0.93.3
+
+### Patch Changes
+
+- [`920e3f5`](https://github.com/lingodotdev/lingo.dev/commit/920e3f5c3ca1fd51b0919db13a4787cfd616de54) Thanks [@mathio](https://github.com/mathio)! - remove cloneDeep for optimization
+
+- Updated dependencies [[`920e3f5`](https://github.com/lingodotdev/lingo.dev/commit/920e3f5c3ca1fd51b0919db13a4787cfd616de54)]:
+  - @lingo.dev/_compiler@0.1.2
+
+## 0.93.2
+
+### Patch Changes
+
+- [`cdb59dd`](https://github.com/lingodotdev/lingo.dev/commit/cdb59dddcd14da1ba3181a33c4c119af877cb4f3) Thanks [@mathio](https://github.com/mathio)! - update deps
+
+## 0.93.1
+
+### Patch Changes
+
+- [`caef325`](https://github.com/lingodotdev/lingo.dev/commit/caef3253bc99fa7bf7a0b40e5604c3590dcb4958) Thanks [@mathio](https://github.com/mathio)! - release fix
+
+- Updated dependencies [[`caef325`](https://github.com/lingodotdev/lingo.dev/commit/caef3253bc99fa7bf7a0b40e5604c3590dcb4958)]:
+  - @lingo.dev/_compiler@0.1.1
+  - @lingo.dev/_react@0.1.1
+
+## 0.93.0
+
+### Minor Changes
+
+- [`e980e84`](https://github.com/lingodotdev/lingo.dev/commit/e980e84178439ad70417d38b425acf9148cfc4b6) Thanks [@maxprilutskiy](https://github.com/maxprilutskiy)! - added the compiler
+
+### Patch Changes
+
+- Updated dependencies [[`e980e84`](https://github.com/lingodotdev/lingo.dev/commit/e980e84178439ad70417d38b425acf9148cfc4b6)]:
+  - @lingo.dev/_compiler@0.1.0
+  - @lingo.dev/_react@0.1.0
+  - @lingo.dev/_sdk@0.8.0
+  - @lingo.dev/_spec@0.34.0
+
+## 0.92.19
+
+### Patch Changes
+
+- [#828](https://github.com/lingodotdev/lingo.dev/pull/828) [`4e9734c`](https://github.com/lingodotdev/lingo.dev/commit/4e9734ce32749caa95703d2b96ba8af6cc83ef94) Thanks [@mathio](https://github.com/mathio)! - preserve order of comments in po files
+
+## 0.92.18
+
+### Patch Changes
+
+- [#824](https://github.com/lingodotdev/lingo.dev/pull/824) [`69d9568`](https://github.com/lingodotdev/lingo.dev/commit/69d9568e85f4f56de3b300b4dd5973bd9c410b99) Thanks [@mathio](https://github.com/mathio)! - skip obsolete entries in po files
+
+## 0.92.17
+
+### Patch Changes
+
+- [#816](https://github.com/lingodotdev/lingo.dev/pull/816) [`28a19e6`](https://github.com/lingodotdev/lingo.dev/commit/28a19e686bc13788b10fd7d9fa6769a34f86d523) Thanks [@vrcprl](https://github.com/vrcprl)! - add config get/set/unset commands
+
 ## 0.92.16
 
 ### Patch Changes